# openapi/okoscope-v1.yaml
openapi: 3.1.0
info:
  title: Okoscope Web API
  version: 1.0.0
  description: Stable API contract for the separately deployed Okoscope Web UI. Timestamps are RFC 3339 UTC values.
servers:
  - url: /
security:
  - sessionAuth: []
paths:
...
  /api/v1/runtime-groups:
    get:
      operationId: listRuntimeGroups
      parameters:
        - { name: project_id, in: query, required: true, schema: { type: string, format: uuid } }
        - {
            name: application_id,
            in: query,
            required: true,
            schema: { type: string, format: uuid },
          }
        - { name: event_kind, in: query, schema: { $ref: '#/components/schemas/RuntimeEventKind' } }
        - {
            name: status,
            in: query,
            schema: { type: string, enum: [open, acknowledged, resolved] },
          }
        - { name: namespace, in: query, schema: { type: string } }
        - { name: workload_kind, in: query, schema: { type: string } }
        - { name: workload_name, in: query, schema: { type: string } }
        - { name: since, in: query, schema: { type: string, format: date-time } }
        - { name: first_seen_from, in: query, schema: { type: string, format: date-time } }
        - { name: first_seen_to, in: query, schema: { type: string, format: date-time } }
        - { name: last_seen_to, in: query, schema: { type: string, format: date-time } }
        - { name: release_id, in: query, schema: { type: string, format: uuid } }
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
      responses: { '200': { $ref: '#/components/responses/RuntimeGroupPage' } }
  /api/v1/runtime-groups/{group_id}:
    parameters: [{ $ref: '#/components/parameters/GroupId' }]
    get:
      operationId: getRuntimeGroup
      responses: { '200': { $ref: '#/components/responses/RuntimeGroupDetail' } }
  /api/v1/runtime-groups/{group_id}/occurrences:
    parameters: [{ $ref: '#/components/parameters/GroupId' }]
    get:
      operationId: listRuntimeGroupOccurrences
      parameters:
        [{ $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }]
      responses:
        {
          '200': { $ref: '#/components/responses/OccurrencePage' },
          '404': { $ref: '#/components/responses/Error' },
        }
  /api/v1/runtime-groups/{group_id}/acknowledge:
    parameters: [{ $ref: '#/components/parameters/GroupId' }]
    post:
      operationId: acknowledgeRuntimeGroup
      responses:
        {
          '200': { $ref: '#/components/responses/RuntimeGroup' },
          '400': { $ref: '#/components/responses/Error' },
          '404': { $ref: '#/components/responses/Error' },
        }
  /api/v1/runtime-groups/{group_id}/resolve:
    parameters: [{ $ref: '#/components/parameters/GroupId' }]
    post:
      operationId: resolveRuntimeGroup
      responses:
        {
          '200': { $ref: '#/components/responses/RuntimeGroup' },
          '400': { $ref: '#/components/responses/Error' },
          '404': { $ref: '#/components/responses/Error' },
        }
  /api/v1/runtime-groups/{group_id}/reopen:
    parameters: [{ $ref: '#/components/parameters/GroupId' }]
    post:
      operationId: reopenRuntimeGroup
      responses:
        {
          '200': { $ref: '#/components/responses/RuntimeGroup' },
          '400': { $ref: '#/components/responses/Error' },
          '404': { $ref: '#/components/responses/Error' },
        }
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: okoscope_session
      description: Opaque HttpOnly browser session; clients send requests with credentials enabled.
    adminAuth: { type: http, scheme: bearer, description: System administrator credential }
  parameters:
    Cursor:
      {
        name: cursor,
        in: query,
        schema: { type: string, format: uuid },
        description: Opaque cursor scoped to the authenticated collection.,
      }
...
    Limit:
      { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
...
    GroupId: { name: group_id, in: path, required: true, schema: { type: string, format: uuid } }
...
  headers:
    RequestId: { schema: { type: string }, description: Correlation identifier for the request. }
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [error, message, request_id]
      properties:
        error: { type: string, example: not_found }
        message: { type: string, example: resource not found }
        request_id: { type: string, example: 0ec02ed2-8483-4981-893e-bffc535897d7 }
        fields:
          type: object
          additionalProperties: { type: string }
          description: Present for field-level validation failures.
...
    RuntimeGroup:
      type: object
      additionalProperties: false
      required:
        [
          id,
          project_id,
          application_id,
          cluster_id,
          namespace,
          workload_kind,
          workload_name,
          fingerprint_version,
          event_kind,
          semantic_summary,
          status,
          first_seen_at,
          first_seen_event_id,
          last_seen_at,
          occurrence_count,
          representative_event_id,
          status_changed_at,
          status_changed_by,
        ]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        project_id: { $ref: '#/components/schemas/Uuid' }
        application_id: { $ref: '#/components/schemas/Uuid' }
        cluster_id: { $ref: '#/components/schemas/Uuid' }
        namespace: { type: string }
        workload_kind: { type: string }
        workload_name: { type: string }
        fingerprint_version: { type: integer }
        event_kind: { type: string }
        semantic_summary: { $ref: '#/components/schemas/RuntimeEventSemanticSummary' }
        status: { type: string, enum: [open, acknowledged, resolved] }
        first_seen_at: { $ref: '#/components/schemas/Timestamp' }
        first_seen_event_id: { $ref: '#/components/schemas/Uuid' }
        last_seen_at: { $ref: '#/components/schemas/Timestamp' }
        occurrence_count: { type: integer, format: int64, minimum: 0 }
        representative_event_id: { $ref: '#/components/schemas/Uuid' }
        status_changed_at: { $ref: '#/components/schemas/NullableTimestamp' }
        status_changed_by: { $ref: '#/components/schemas/NullableUuid' }
    RuntimeGroupPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/RuntimeGroup' } }
        next_cursor: { $ref: '#/components/schemas/NullableUuid' }
    EventOccurrence:
      type: object
      additionalProperties: false
      required:
        [
          id,
          event_id,
          observed_at,
          received_at,
          node_name,
          namespace,
          pod_name,
          container_name,
          process_command,
          event_kind,
          payload,
          correlation,
          related_evidence,
          release_id,
          release_version,
        ]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        event_id: { $ref: '#/components/schemas/Uuid' }
        observed_at: { $ref: '#/components/schemas/Timestamp' }
        received_at: { $ref: '#/components/schemas/Timestamp' }
        node_name: { type: string }
        namespace: { type: string }
        pod_name: { type: string }
        container_name: { type: string }
        process_command: { type: string }
        event_kind: { type: string }
        payload: { $ref: '#/components/schemas/RuntimeEventPayload' }
        correlation: { $ref: '#/components/schemas/EventCorrelation' }
        related_evidence: { type: array, maxItems: 20, items: { $ref: '#/components/schemas/RelatedEvidence' } }
        release_id: { $ref: '#/components/schemas/NullableUuid' }
        release_version: { type: [string, 'null'] }
    OccurrencePage:
      type: object
      additionalProperties: false
      required: [items, next_cursor, ordering]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/EventOccurrence' } }
        next_cursor: { $ref: '#/components/schemas/NullableUuid' }
        ordering: { type: string, const: received_at_desc_observed_at_desc_id_desc }
...
    RuntimeEventSemanticSummary:
      oneOf:
        - { $ref: '#/components/schemas/ProcessExecSemanticSummary' }
        - { $ref: '#/components/schemas/SyscallSemanticSummary' }
        - { $ref: '#/components/schemas/NetworkConnectSemanticSummary' }
        - { $ref: '#/components/schemas/InboundNetworkSemanticSummary' }
        - { $ref: '#/components/schemas/NetworkDnsQuerySemanticSummary' }
        - { $ref: '#/components/schemas/NetworkDnsResponseSemanticSummary' }
        - { $ref: '#/components/schemas/FileActivitySemanticSummary' }
        - { $ref: '#/components/schemas/ProcessExitSemanticSummary' }
        - { $ref: '#/components/schemas/ContainerTerminationSemanticSummary' }
        - { $ref: '#/components/schemas/ContainerRestartSemanticSummary' }
        - { $ref: '#/components/schemas/RestartLoopSemanticSummary' }
    ProcessExecSemanticSummary:
      type: object
      additionalProperties: false
      required: [executable]
      properties:
        executable: { type: string, minLength: 1 }
...
    RuntimeEventKind:
      type: string
      enum: [process.exec, syscall, network.connect, network.listen, network.accept, network.dns_query, network.dns_response, file.create, file.modify, file.delete, file.rename, process.exit, container.terminated, container.restart, container.restart_loop]
...
    RuntimeEventPayload:
      oneOf:
        - { $ref: '#/components/schemas/ProcessExecPayload' }
        - { $ref: '#/components/schemas/SyscallPayload' }
        - { $ref: '#/components/schemas/NetworkConnectPayload' }
        - { $ref: '#/components/schemas/NetworkListenPayload' }
        - { $ref: '#/components/schemas/NetworkAcceptPayload' }
        - { $ref: '#/components/schemas/NetworkDnsQueryPayload' }
        - { $ref: '#/components/schemas/NetworkDnsResponsePayload' }
        - { $ref: '#/components/schemas/FileCreatePayload' }
        - { $ref: '#/components/schemas/FileModifyPayload' }
        - { $ref: '#/components/schemas/FileDeletePayload' }
        - { $ref: '#/components/schemas/FileRenamePayload' }
        - { $ref: '#/components/schemas/ProcessExitPayload' }
        - { $ref: '#/components/schemas/ContainerTerminationPayload' }
        - { $ref: '#/components/schemas/ContainerRestartPayload' }
        - { $ref: '#/components/schemas/ContainerRestartLoopPayload' }
...
    FirstSeenNotificationSummary:
      type: object
      additionalProperties: false
      required: [state, delivery_count, succeeded_count, failed_count]
      properties:
        state:
          {
            type: string,
            enum:
              [
                pending,
                not_configured,
                delivering,
                delivered,
                terminally_failed,
                backfill_suppressed,
              ],
          }
        delivery_count: { type: integer, format: int64, minimum: 0 }
        succeeded_count: { type: integer, format: int64, minimum: 0 }
        failed_count: { type: integer, format: int64, minimum: 0 }
    RuntimeGroupDetail:
      unevaluatedProperties: false
      allOf:
        - { $ref: '#/components/schemas/RuntimeGroup' }
        - type: object
          required: [representative_event, notification]
          properties:
            representative_event: { $ref: '#/components/schemas/EventOccurrence' }
            notification: { $ref: '#/components/schemas/FirstSeenNotificationSummary' }
...
    Uuid: { type: string, format: uuid }
    NullableUuid: { type: [string, 'null'], format: uuid }
...
    Timestamp: { type: string, format: date-time, example: '2026-08-17T12:00:00Z' }
    NullableTimestamp: { type: [string, 'null'], format: date-time }
...
  responses:
...
    RuntimeGroupPage:
      {
        description: Runtime group page,
        headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } },
        content:
          { application/json: { schema: { $ref: '#/components/schemas/RuntimeGroupPage' } } },
      }
    RuntimeGroup:
      {
        description: Runtime group,
        headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } },
        content: { application/json: { schema: { $ref: '#/components/schemas/RuntimeGroup' } } },
      }
    RuntimeGroupDetail:
      {
        description: Runtime group detail,
        headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } },
        content:
          { application/json: { schema: { $ref: '#/components/schemas/RuntimeGroupDetail' } } },
      }
    OccurrencePage:
      {
        description: Runtime group occurrence page,
        headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } },
        content: { application/json: { schema: { $ref: '#/components/schemas/OccurrencePage' } } },
      }
...
    Error:
      description: Correlated API error
      headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }